Fixup code to comply with more modern standards.
