+++ src/windows/ScreenTime.csproj
@@ -1,0 +1,35 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>WinExe</OutputType>
+    <!-- <OutputType>Exe</OutputType> -->
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <TargetFramework>net9.0-windows</TargetFramework>
+    <UseWindowsForms>true</UseWindowsForms>
+    <Description>this will monitor your screentime and put it in the taskbar. It will lock you out when you've exceeded your time plus a grace period.</Description>
+    <Title>ScreenTime</Title>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Humanizer.Core" Version="2.14.1" />
+    <PackageReference Include="Microsoft.Extensions.Hosting" Version="9.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Http" Version="9.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Http.Resilience" Version="9.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Logging" Version="9.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Configuration" Version="9.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="9.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="9.0.0" />
+    <PackageReference Include="Microsoft.Win32.SystemEvents" Version="9.0.0" />
+    <PackageReference Include="NetEscapades.Extensions.Logging.RollingFile" Version="2.5.0" />
+    <PackageReference Include="System.DirectoryServices.AccountManagement" Version="9.0.0" />
+    <PackageReference Include="System.Management" Version="9.0.0" />
+    <!-- PackageReference Include="Microsoft.WindowsAppSDK" Version="1.6.241114003" / -->
+  </ItemGroup>
+  <Target Name="ZipOutput" AfterTargets="Build">
+    <ItemGroup>
+      <FilesToZip Include="$(OutputPath)**\*.*" />
+    </ItemGroup>
+    <!-- Exec Command="powershell Compress-Archive -Path $(OutputPath) -DestinationPath $(OutputPath)screentime.zip" / -->
+  </Target>
+</Project>
